+++ app/planning.py
@@ -1,0 +1,3 @@
+"""OCR previews cannot become executable plans in this milestone."""
+def build_plan(*args, **kwargs):
+    raise NotImplementedError('Preparation planning is not enabled in milestone 1')